/trunk: cmnRandomSequence updated to support int, long, long long.  Added specific methods for size_t and ptrdiff_t.  Use conditional compilation to check data model (LLP64, LP64) to generate 64 bits random ints.  Updated cisstVector tests to avoid many warnings with MS compiler 64 bits and use new cmnRandomSequence methods when needed.  cisstNumerical tests updated to use cmnRandomSequence methods.  nmrSVDEcomomy fixed to use cisstNetlib 3.0.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3927 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstVector/tests/vctDataFunctionsFixedSizeMatrixTest.cpp
@@ -146,7 +146,7 @@ void vctDataFunctionsFixedSizeMatrixTest::TestScalar(void)
     position = 0;
     for (row = 0; row < mInt.rows(); ++row) {
         for (col = 0; col < mInt.cols(); ++col) {
-            mInt.Element(row, col) = row * 10 + col;
+            mInt.Element(row, col) = static_cast<int>(row * 10 + col);
             CPPUNIT_ASSERT_EQUAL(static_cast<double>(row * 10 + col), cmnDataScalar(mInt, position));
             position++;
         }
@@ -157,7 +157,7 @@ void vctDataFunctionsFixedSizeMatrixTest::TestScalar(void)
         for (col = 0; col < mmDouble.cols(); ++col) {
             for (subRow = 0; subRow < mmDouble.Element(row, col).rows(); ++subRow) {
                 for (subCol = 0; subCol < mmDouble.Element(row, col).cols(); ++subCol) {
-                    mmDouble.Element(row, col).Element(subRow, subCol) = row * col * 100 + subRow * subCol;
+                    mmDouble.Element(row, col).Element(subRow, subCol) = static_cast<double>(row * col * 100 + subRow * subCol);
                     CPPUNIT_ASSERT_EQUAL(static_cast<double>(row * col * 100 + subRow * subCol),
                                          cmnDataScalar(mmDouble, position));
                     position++;
@@ -172,7 +172,7 @@ void vctDataFunctionsFixedSizeMatrixTest::TestScalar(void)
         for (col = 0; col < mmMixed.cols(); ++col) {
             for (subRow = 0; subRow < mmMixed.Element(row, col).rows(); ++subRow) {
                 for (subCol = 0; subCol < mmMixed.Element(row, col).cols(); ++subCol) {
-                    mmMixed.Element(row, col).Element(subRow, subCol) = row * col * 100 + subRow * subCol;
+                    mmMixed.Element(row, col).Element(subRow, subCol) = static_cast<double>(row * col * 100 + subRow * subCol);
                     CPPUNIT_ASSERT_EQUAL(static_cast<double>(row * col * 100 + subRow * subCol),
                                          cmnDataScalar(mmMixed, position));
                     position++;
